+++ bin/ons_csv_to_ctb_json_load.py
@@ -12,7 +12,7 @@ PUBLIC_SECURITY_MNEMONIC = 'PUB'
 GEOGRAPHIC_VARIABLE_TYPE = 'GEOG'
 
 DatabaseClassifications = namedtuple('DatabaseClassifications',
-                                     'classifications lowest_geog_variable')
+                                     'classifications lowest_geog_variable non_public')
 
 
 def isnumeric(string):
@@ -61,10 +61,10 @@ class Loader:
     Many of the fields in this class are cached properties, with the data loaded on first access.
     """
 
-    def __init__(self, input_directory, geography_file, best_effort=False, dataset_filter=''):
+    def __init__(self, input_directory, geography_files, best_effort=False, dataset_filter=''):
         """Initialise MetadataLoader object."""
         self.input_directory = input_directory
-        self.geography_file = geography_file
+        self.geography_files = geography_files
         self.dataset_filter = dataset_filter
         self._error_count = 0
 
@@ -411,8 +411,8 @@ class Loader:
 
             database_mnemonic = database.pop('Database_Mnemonic')
 
-            db_classifications = database_to_classifications.get(database_mnemonic,
-                                                                 DatabaseClassifications([], None))
+            db_classifications = database_to_classifications.get(
+                database_mnemonic, DatabaseClassifications([], None, []))
             database['Lowest_Geog_Variable'] = db_classifications.lowest_geog_variable
 
             database_type_code = database.pop('Database_Type_Code')
@@ -427,7 +427,8 @@ class Loader:
                              database.pop('Database_Description'),
                              database.pop('Database_Description_Welsh')),
                          'Classifications': db_classifications.classifications,
-                         'Database_Type_Code': database_type_code})
+                         'Database_Type_Code': database_type_code,
+                         'Non_Public_Classifications': db_classifications.non_public})
 
         return databases
 
@@ -501,23 +502,30 @@ class Loader:
                                                             default_to_english=False)
 
         # Categories for geographic variables are supplied in a separate file.
-        if not self.geography_file:
-            logging.info('No geography file specified')
+        if not self.geography_files:
+            logging.info('No geography files specified')
             return categories
 
-        for class_name, geo_cats in read_geo_cats(self.geography_file).items():
-            if class_name not in self.classifications:
-                logging.info(f'Reading {self.geography_file}: found Welsh labels for unknown '
-                             f'classification: {class_name}')
+        # read_geo_cats returns a dictionary of lower case variable name to categories.
+        # This allows the reader to be case agnostic with regards to column headings.
+        for lc_name, geo_cats in read_geo_cats(self.geography_files).items():
+            for name in self.classifications:
+                if name.lower() == lc_name:
+                    class_name = name
+                    break
+            else:
+                logging.info(f'Reading {geo_cats.source_file}: found labels for unknown '
+                             f'geographic classification: {lc_name}')
                 continue
 
             if not self.classifications[class_name].private['Is_Geographic']:
-                self.recoverable_error(f'Reading {self.geography_file}: found Welsh labels for '
+                self.recoverable_error(f'Reading {geo_cats.source_file}: found labels for '
                                        f'non geographic classification: {class_name}')
                 continue
 
-            english_names = {cd: nm.name for cd, nm in geo_cats.items() if nm.name}
-            welsh_names = {cd: nm.welsh_name for cd, nm in geo_cats.items() if nm.welsh_name}
+            english_names = {cd: nm.name for cd, nm in geo_cats.code_to_label.items() if nm.name}
+            welsh_names = {cd: nm.welsh_name for cd, nm in geo_cats.code_to_label.items()
+                           if nm.welsh_name}
 
             # If Welsh names are not provided then do not substitute English names in their place.
             # The Welsh base dataset includes the base English dataset and the metadata server
@@ -525,6 +533,14 @@ class Loader:
             categories[class_name] = Bilingual(english_names,
                                                welsh_names if welsh_names else None,
                                                default_to_english=False)
+
+        geos_with_labels = []
+        for class_name, classification in self.classifications.items():
+            if classification.private['Is_Geographic'] and class_name in categories:
+                geos_with_labels.append(class_name)
+        if geos_with_labels:
+            logging.info('Labels supplied for these geographic classifications: '
+                         f'{sorted(geos_with_labels)}')
 
         return categories
 
@@ -827,6 +843,10 @@ class Loader:
             num_cat_items = classification.pop('Number_Of_Category_Items')
             num_cat_items = int(num_cat_items) if num_cat_items else 0
 
+            # The Cantabular_Public_Flag is set for each classification on a per database basis.
+            # The base version has the flag set to Y, and may be set to N for other databases.
+            classification['Cantabular_Public_Flag'] = 'Y'
+
             classifications[classification_mnemonic] = BilingualDict(
                 classification,
                 # The private fields and not included in the English/Welsh variants of datasets.
@@ -854,6 +874,7 @@ class Loader:
                 classifications[variable_mnemonic] = BilingualDict(
                     {
                         'Mnemonic_2011': None,
+                        'Cantabular_Public_Flag': 'Y',
                         'Default_Classification_Flag': None,
                         'Version': variable.private['Version'],
                         'ONS_Variable': variable,
@@ -979,6 +1000,7 @@ class Loader:
         for database_mnemonic, db_vars in db_to_raw_vars.items():
             lowest_geog_var = None
             classifications = set()
+            non_public = set()
             contains_geo_vars = False
             for db_var in db_vars:
                 database_mnemonic = db_var['Database_Mnemonic']
@@ -1005,6 +1027,7 @@ class Loader:
 
                 # Add the specific classification to the database if Classification_Mnemonic is set
                 # else add all the classifications for the variable.
+                row_classifications = set()
                 if classification_mnemonic:
                     if classification_mnemonic not in \
                             variable_to_classifications.get(variable_mnemonic, set()):
@@ -1013,10 +1036,15 @@ class Loader:
                             f'{classification_mnemonic} is unknown Classification_Mnemonic for '
                             f'Variable_Mnemonic {variable_mnemonic}')
                     else:
+                        row_classifications = {classification_mnemonic}
                         classifications.add(classification_mnemonic)
                 else:
-                    classifications = classifications.union(variable_to_classifications.get(
-                        variable_mnemonic, set()))
+                    row_classifications = variable_to_classifications.get(
+                        variable_mnemonic, set())
+                    classifications = classifications.union(row_classifications)
+
+                if db_var['Cantabular_Public_Flag'] == 'N':
+                    non_public.update(row_classifications)
 
             if not lowest_geog_var and contains_geo_vars:
                 self.recoverable_error(f'Reading {self.full_filename(filename)} '
@@ -1024,7 +1052,8 @@ class Loader:
                                        f'variable for database {database_mnemonic}')
 
             database_to_classifications[database_mnemonic] = DatabaseClassifications(
-                classifications=classifications, lowest_geog_variable=lowest_geog_var)
+                classifications=classifications, lowest_geog_variable=lowest_geog_var,
+                non_public=sorted(list(non_public)))
 
         return database_to_classifications